Create README.md
Part of code written while working in Institue of hman genetics in Heidelberg. Main Project there was pipeline for snv variant calling for data from WES. Beside the pipeline a data base was designed for collecting variants and the upload scripts. Thee project was not accomplished, would need some improvements and the code is not very clean, but it was tested and it does what should do.